Fix PID algorithm: increase Kp, expand integral band, and migrate temperature to double
- PID tuning: Kp 8->20 (stronger P response), IntegralBand 5->15 (allows I accumulation far from target)
- Root cause: at 85C with target 75C, |error|=10 exceeded IntegralBand=5 so I was always reset, leaving only P=80W correction - insufficient under heavy load
- With fix: P=200W immediate correction, I accumulates progressively reducing power down to 275W
- Temperature type system: changed all temperature values from uint to double for precision
  - IGpuDevice.GetTemperature(out double), SimulatedGpuDevice, NvmlGpuDevice
  - ThermalControllerEventArgs, MetricsSnapshot, DashboardEvent
  - TriggerEvaluator.Evaluate(double), ConsoleDashboard.CreateTempBar(double)
  - MockGpuDevice temperature methods and test data
- Dashboard: temperature displayed with F1 format (e.g. '85.0C')
- Tests: all 118 tests pass

+++ Dashboard/DashboardEvent.cs
@@ -8,7 +8,7 @@ namespace GpuThermalController.Dashboard;
 /// </summary>
 public record DashboardEvent(
     DateTime Timestamp,
-    uint Temperature,
+    double Temperature,
     int PowerLimit,
     bool IsControlling,
     ControllerEventType EventType,
